[Feature] Introduce `benchx_cli` for benchmark
Use it with `./benchx_cli -o <output_ptrace_file> run <lynx_bundle_path>` to get a ptrace file of the Lynx Bundle.

Under the hood it runs the Lynx Bundle with specific embedder api mocked so not suitable for serious e2e test, but should be sufficient for benchmark.

Also, codspeed is integrate into the cli, so benchmark result can be uploaded to codspeed platform for tracking ingression.